Extract common Contract/Invoice row behaviour
Create Export::SubmissionEntryRow, which is responsible for looking
up mapped fields for us and also defining the 8 AdditionalN fields
common to Contract/Invoice
